Move Decision table logic all in one package
