+++ src/main/scala/com/gshakhn/idea/idea/fitnesse/decisiontable/DecisionOutput.scala
@@ -1,0 +1,8 @@
+package com.gshakhn.idea.idea.fitnesse.decisiontable
+
+import com.gshakhn.idea.idea.fitnesse.lang.psi.Cell
+import com.intellij.lang.ASTNode
+
+class DecisionOutput(node: ASTNode) extends Cell(node) {
+
+}
