+++ src/main/scala/com/gshakhn/idea/idea/fitnesse/decisiontable/DecisionTable.scala
@@ -1,0 +1,8 @@
+package com.gshakhn.idea.idea.fitnesse.decisiontable
+
+import com.gshakhn.idea.idea.fitnesse.lang.psi.Table
+import com.intellij.lang.ASTNode
+
+class DecisionTable(node: ASTNode) extends Table(node) {
+
+}
