--- src/main/scala/com/gshakhn/idea/idea/fitnesse/lang/psi/DecisionOutput.scala
@@ -1,7 +1,0 @@
-package com.gshakhn.idea.idea.fitnesse.lang.psi
-
-import com.intellij.lang.ASTNode
-
-class DecisionOutput(node: ASTNode) extends Cell(node) {
-
-}
