--- src/main/scala/com/gshakhn/idea/idea/fitnesse/lang/psi/DecisionTable.scala
@@ -1,7 +1,0 @@
-package com.gshakhn.idea.idea.fitnesse.lang.psi
-
-import com.intellij.lang.ASTNode
-
-class DecisionTable(node: ASTNode) extends Table(node) {
-
-}
